refactor: remove leftover module alias glue
